+++ dem_stitcher/merge.py
@@ -13,12 +13,13 @@ from .rio_window import get_indices_from_extent
 def merge_tile_datasets(datasets: List[rasterio.DatasetReader],
                         bounds: list = None,
                         resampling: str = 'nearest',
-                        res_buffer: int = 0
+                        res_buffer: int = 0,
+                        nodata: float = np.nan
                         ) -> Tuple[np.ndarray, dict]:
     merged_arr, merged_transform = merge(datasets,
                                          resampling=Resampling[resampling],
                                          # This fixes the nodata values
-                                         nodata=np.nan,
+                                         nodata=nodata,
                                          # This fixes the float32 output
                                          dtype='float32',
                                          )