+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Clarity about driver keyword in `stitch_dem` in readme, docstrings
 - Ensure overlap of tiles is non-trivial AND polygonal (excludes point and line intersections)
 - Similar check of polygonal type for window reading for better error handling
+- Add `merge_nodata_value` to `merge_tile_datasets`, `merge_and_transform_dem_tiles`, and `stitch_dem` to allow for fill value of 0. As such, nodata areas within DEM tiles when converted to Ellipsoidal height will be filled in with geoid values. No other values outside of `np.nan` or `0` permitted.
+
 
 ### Changed
 - Moved functions into more logical python file including merge calls into `merge.py` and tile functions into `datasets.py`